api/requirements.txt
aiofiles==0.5.0
aiohttp==3.7.4.post0
aniso8601==7.0.0
asgiref==3.4.1
async-exit-stack==1.0.1
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.5.7
backports.entry-points-selectable==1.1.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.15.0
chardet==4.0.0
click==7.1.2
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
cryptography==36.0.1
distlib==0.3.2
Django==3.2.9
dnspython==2.2.0
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.68.1
filelock==3.0.12
Flask==2.0.1
googlemaps==4.5.3
graphene==2.1.9
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet==1.1.1
h11==0.12.0
httptools==0.1.2
idna==2.10
importlib-metadata==4.8.1
itsdangerous==1.1.0
Jinja2==2.11.3
lxml==4.6.3
MarkupSafe==2.0.1
multidict==5.2.0
numpy==1.21.2
orjson==3.6.7
pandas==1.3.3
passlib==1.7.4
pipenv==2021.5.29
platformdirs==2.3.0
promise==2.3
psycopg2==2.9.2
psycopg2-binary==2.9.3
pyasn1==0.4.8
pycodestyle==2.7.0
pycparser==2.21
pydantic==1.8.2
python-dateutil==2.8.2
python-dotenv==0.19.2
python-jose==3.3.0
python-multipart==0.0.5
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.8
Rx==1.6.1
selenium==3.141.0
six==1.16.0
soupsieve==2.2.1
SQLAlchemy==1.4.25
sqlparse==0.4.2
starlette==0.14.2
toml==0.10.2
typing-extensions==3.10.0.2
ujson==4.3.0
urllib3==1.26.4
uvicorn==0.13.4
uvloop==0.16.0
virtualenv==20.7.2
virtualenv-clone==0.5.7
watchgod==0.7
webdriver-manager==3.4.2
websockets==8.1
Werkzeug==2.0.1
yarl==1.7.0
zipp==3.5.0
